checking failiure

+++ libsdrt/src/d/rt/unwind.d
@@ -1,7 +1,7 @@
 module d.rt.unwind;
 
 alias _Unwind_Ptr = void*;
-alias _Unwind_Word = size_t;
+alias _Unwind_Word = ulong;
 
 extern(C):
